leeey-005 1. 모듈 구조 변경 2. infra module
