+++ buildSrc/src/main/groovy/project.laflex.java-conventions.gradle
@@ -9,20 +9,14 @@ repositories {
 }
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-webflux'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'io.projectreactor:reactor-test'
-
     testImplementation 'org.spockframework:spock-core:2.3-groovy-4.0'
     testImplementation 'org.spockframework:spock-spring:2.3-groovy-4.0'
-
-    implementation 'org.springdoc:springdoc-openapi-starter-webflux-ui:2.2.0'
+    testImplementation 'org.instancio:instancio-core:3.3.0'
 
     compileOnly 'org.projectlombok:lombok'
     implementation 'org.springframework.boot:spring-boot-devtools'
     annotationProcessor 'org.projectlombok:lombok'
-
-    implementation 'org.springdoc:springdoc-openapi-starter-webflux-ui:2.2.0'
 
     implementation 'org.hibernate.validator:hibernate-validator:8.0.1.Final'
     implementation 'org.glassfish:jakarta.el:4.0.2'
