--- buildSrc/src/main/groovy/project.laflex.library-conventions.gradle
@@ -1,7 +1,0 @@
-plugins {
-    id 'java-library'
-    id 'maven-publish'
-    id 'project.laflex.java-conventions'
-}
-
-group = 'project.laflex'
