+++ buildSrc/src/main/groovy/project.laflex.library-webflux-conventions.gradle
@@ -1,0 +1,13 @@
+plugins {
+    id 'java-library'
+    id 'maven-publish'
+    id 'project.laflex.java-conventions'
+}
+
+group = 'project.laflex'
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-webflux'
+    implementation 'org.springdoc:springdoc-openapi-starter-webflux-ui:2.2.0'
+    testImplementation 'io.projectreactor:reactor-test'
+}
