+++ common/src/main/java/project/laflex/common/FailureResponse.java
@@ -1,0 +1,14 @@
+package project.laflex.common;
+
+import java.util.List;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class FailureResponse<T> {
+
+  private Long timestamp;
+
+  private List<T> errors;
+}
