+++ infra/build.gradle
@@ -1,10 +1,12 @@
 plugins {
     id 'groovy'
     id 'project.laflex.library-mvc-conventions'
-    id 'project.laflex.library-webflux-conventions'
 }
 
 dependencies {
+    implementation project(':common')
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    runtimeOnly 'com.h2database:h2:2.2.224'
 }
 
 test {
