+++ infra/src/main/java/project/laflex/infra/InfraApplication.java
@@ -1,0 +1,13 @@
+package project.laflex.infra;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = "project.laflex.infra.*")
+public class InfraApplication {
+
+  public static void main(String[] args) {
+
+    SpringApplication.run(InfraApplication.class, args);
+  }
+}
