+++ infra/src/main/resources/application.yml
@@ -1,0 +1,34 @@
+server:
+  port: 8082
+  error:
+    include-stacktrace: never
+spring:
+  application:
+    name: laflex-infra
+  devtools:
+    add-properties: false
+  datasource:
+    url: jdbc:h2:mem:testdb;MODE=MYSQL
+    driver-class-name: org.h2.Driver
+    username: sa
+    password:
+  jpa:
+    database: h2
+    database-platform: org.hibernate.dialect.H2Dialect
+    defer-datasource-initialization: true
+    properties:
+      hibernate:
+        show_sql: true
+        format_sql: true
+    hibernate:
+      ddl-auto: create-drop
+  h2:
+    console:
+      enabled: true
+      path: /h2-console
+      settings:
+        trace: false
+        web-allow-others: false
+  sql:
+    init:
+      mode: always
