--- lotto/build.gradle
@@ -1,13 +1,0 @@
-plugins {
-    id 'groovy'
-    id 'project.laflex.library-conventions'
-}
-
-dependencies {
-    implementation project(':common')
-    implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-csv:2.15.2'
-}
-
-test {
-    useJUnitPlatform()
-}
