--- lotto/src/main/java/project/laflex/lotto/config/SuccessResponse.java
@@ -1,13 +1,0 @@
-package project.laflex.lotto.config;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class SuccessResponse<T> {
-
-  private Long timestamp;
-
-  private T result;
-}
