--- lotto/src/main/java/project/laflex/lotto/dto/NumberCountDto.java
@@ -1,13 +1,0 @@
-package project.laflex.lotto.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class NumberCountDto {
-
-  private Integer number;
-
-  private Integer count;
-}
