--- lotto/src/main/resources/application.yml
@@ -1,12 +1,0 @@
-spring:
-  application:
-    name: laflex
-server:
-  error:
-    include-stacktrace: never
-
-springdoc:
-  api-docs:
-    path: /api-docs
-  swagger-ui:
-    path: /swagger-ui.html
