--- lotto/src/test/groovy/project/laflex/lotto/util/LottoHistoryUtilTest.groovy
@@ -1,16 +1,0 @@
-package project.laflex.lotto.util
-
-
-import spock.lang.Specification
-
-class LottoHistoryUtilTest extends Specification {
-
-    def "getHistories"() {
-
-        when:
-        def histories = HistoryUtil.getHistories()
-
-        then:
-        histories.size() == 1086
-    }
-}
