+++ webflux/src/main/java/project/laflex/webflux/app/WebFluxApplication.java
@@ -1,0 +1,15 @@
+package project.laflex.webflux.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.web.reactive.config.EnableWebFlux;
+
+@EnableWebFlux
+@SpringBootApplication(scanBasePackages = "project.laflex.webflux.*")
+public class WebFluxApplication {
+
+  public static void main(String[] args) {
+
+    SpringApplication.run(WebFluxApplication.class, args);
+  }
+}
