+++ webflux/src/main/java/project/laflex/webflux/app/lotto/dto/LottoMostNumbersOfYearDto.java
@@ -1,0 +1,14 @@
+package project.laflex.webflux.app.lotto.dto;
+
+import java.util.List;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class LottoMostNumbersOfYearDto {
+
+  private Integer year;
+
+  private List<Integer> numbers;
+}
