+++ webflux/src/main/java/project/laflex/webflux/app/lotto/dto/LottoNumberCountDto.java
@@ -1,0 +1,13 @@
+package project.laflex.webflux.app.lotto.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class LottoNumberCountDto {
+
+  private Integer number;
+
+  private Integer count;
+}
