+++ webflux/src/main/java/project/laflex/webflux/app/lotto/handler/LottoHealthCheckHandlerFunction.java
@@ -1,0 +1,21 @@
+package project.laflex.webflux.app.lotto.handler;
+
+import java.util.Map;
+import project.laflex.webflux.config.ResponseWrapper;
+import org.springframework.lang.NonNull;
+import org.springframework.stereotype.Component;
+import org.springframework.web.reactive.function.server.HandlerFunction;
+import org.springframework.web.reactive.function.server.ServerRequest;
+import org.springframework.web.reactive.function.server.ServerResponse;
+import reactor.core.publisher.Mono;
+
+@Component
+public class LottoHealthCheckHandlerFunction implements HandlerFunction<ServerResponse> {
+
+  @NonNull
+  @Override
+  public Mono<ServerResponse> handle(@NonNull ServerRequest request) {
+
+    return ResponseWrapper.success(Map.of("status", "OK"));
+  }
+}
