+++ webflux/src/main/resources/application.yml
@@ -1,0 +1,14 @@
+server:
+  port: 8081
+  error:
+    include-stacktrace: never
+spring:
+  application:
+    name: laflex-webflux
+  devtools:
+    add-properties: false
+springdoc:
+  api-docs:
+    path: /api-docs
+  swagger-ui:
+    path: /swagger-ui.html
